Graph v0.1.0 Graph

Documentation for Graph

This module implements a generic graph data structure, which is used to support the algorithms implemented in Graph.Algorithm. However, it can also serve as a general API for graph representation.

Both directed and undirected graphs are supported, and in the future, functionality specific to weighted graphs will be added.

Struct fields:

The graph state (in terms of nodes and edges) is encapsulated entirely in the “nodes” field, which is a Map.t that maps node IDs to Graph.Node.t structs (see Graph.Node.t for more detail).

The only other fields in Graph.t are the boolean flags :directed and :weighted, which as the names imply, determine the type of graph representation. Unless otherwise specified, a new Graph.t struct is a simple graph (i.e. unweighted and undirected) [note: the current implementation does not prevent a client from adding loops].

To alternatively create a graph of a different type, supply any desired combination of corresponding truth values for :directed and :weighted upon creation. Whenever applicable, the API enforces that functions called are well-defined for the type of graph passed to them as an argument. e.g. you cannot add a directed edge to a Graph.t struct that has :directed set to false.

Types

id()

id() :: binary() | integer()

nodesmap()

nodesmap() :: Map.t(id(), Graph.Node.t())

nodesmap(struct)

nodesmap(struct) :: Map.t(id(), Graph.Node.t(struct))

t()

t() :: %Graph{directed: term(), nodes: nodesmap(), weighted: term()}

t(struct)

t(struct) :: %Graph{directed: term(), nodes: nodesmap(struct), weighted: term()}

Functions

add_edge(graph, sourceNodeID, destNodeID)

Adds an edge to the graph

To succesfully add an edge, add_edge requires that both nodes already exist in the graph.

When the inputs are valid, returns the updated graph

Otherwise, prints an error message

add_node(graph, node)

add_node(Graph.t(), Graph.Node.t()) :: Graph.t()

Adds a node to the graph, as long as a node with that id does not already exist

If the graph already has a node with the id passed to add_node, add_node prints an error and returns the unaltered graph.

get_node(graph, nodeID)

Returns the node struct for the node with the specified id, if it exists, else nil

get_nodes(graph)

Returns a map of ids to graph nodes

has_node?(graph, nodeID)

Returns true i.f.f. graph has a node with id nodeID

new()

Creates an unweighted, undirected graph struct, with an empty nodes map.

new(nodeset)

Creates a new graph struct for an unweighted/undirected graph from “nodeset”, an Enum of Node structs

num_nodes(graph)

Returns the number of nodes in the graph

remove_edge(g, source, dest)

Returns graph with the specified undirected edge removed.

If the edge is not present, prints an error message, and returns nil.

remove_node(graph, nodeID)

remove_node(Graph.t(), binary()) :: Graph.t()

Removes the specified node from the graph, if it exists. if the graph is undirected, also removes any edges connected to it

Returns the updated graph if the node exists, otherwise, nil.

set_node(graph, node)

Sets the key node.id in the graph’s nodes map to the given node