Graph v0.1.0 Graph.Algorithm

Graph.Algorithm provides general algorithms for searching graphs, checking their properties, and so on.

Every function takes a Graph.t struct as its first parameter. Additionally, many of the algorithms involve specific nodes as inputs, and thus take node ids in subsequent parameters.

Functions

bfs_source(graph, sourceNodeID, goalNodeID, opts \\ [])

Performs a BFS search for a particular node, starting from a “source” node.

Requires that a node with the given source id exists in the graph, otherwise, an ArgumentError is raised

Invoke bfs_source(graph, sourceNodeID, goalNodeID) to execute a BFS that begins at the source node and searches for the goal node, and returns a boolean value indicating whether or not the goal is reachable from the source.

Optionally, provide a report_dist: true option for the last parameter, e.g. bfs_source(graph, source, goal, report_dist: true). When report_dist: true and goal is reachable, bfs_source returns a 2-tuple of the format {true, numEdges}, where numEdges is the number of edges in the shortest path (in the unweighted sense) from the source to the goal. If the goal is not reachable, bfs_source always returns false on valid inputs.

dfs_post_order(graph, visitFunc, state)

Runs DFS on the entire graph (as opposed to just from a specific starting node), visiting nodes in postorder while applying a state transition function. Returns the state once every node in the graph has been visited.

Call dfs_post_order/3 like so: dfs_post_order(graph, visitFunc, state)

Takes parameters “state” and an anonymous state transition function “visitFunc”, which together are used to update the state as follows every time a node is visited:

state = visitFunc.(state, node)

The “state” argument passed to dfs_post_order/3 is some arbitrary initial state, with the value depending on the particular application of dfs_post_order/3.

To illustrate, toplogical_sort utilizes dfs_post_order/3 like so: top_sort = dfs_post_order(some_dag, &track_reverse_post_order/2, []).

The initial state is an empty list. track_reverse_post_order is captured and passed as the visitFunc, which takes inputs of the current list (i.e. the “state”) and current node being visited.

Every time a node is visited, the following line is effectively executed: traversal_list = track_reverse_post_order.(traversal_list, node)

track_reverse_post_order/2 prepends “node” to the front of traversal_list state, and returns the new list.

For a DAG, the reverse ordering of a postorder DFS traversal represents a valid topological sort of the graph, so topological_sort can in turn just return the list produced by this call to dfs_post_order/3.

dfs_source?(graph, sourceNodeID, goalNodeID)

Executes a DFS for a goal node from a specific “source” node

Returns a boolean value indicating whether or not the goal is reachable from the starting node

Raises an error if the source node does not exist in the graph

Call as follows: dfs_source?(graph, sourceNodeID, goalNodeID)

is_acyclic?(graph)

Returns true i.f.f a directed graph is acyclic.

Call as follows: is_acyclic?(graph)

This function is undefined for undirected graphs.

topological_sort(g)

For a directed acyclic input graph (DAG), returns a list of node ids representing a toplogical sort of the graph

Returns nil if input graph is not a DAG, in which case topological sorting is undefined

Call as follows: topological_sort(graph)

Note that, in general, a DAG may have multiple valid topological sorts. topological_sort/1 returns one of them. If there are multiple valid sorts, there is no guarantee as to which particular one is returned.