Graph v0.1.0 Graph.Node

Documentation for Graph.Node.

Graph.Node implements a struct to hold the state of a graph node. The Graph module builds on Graph.Node to represent graphs out of Node.t structs.

Struct fields: The “id” (generally a binary or an integer) field uniquely identifies a node stored a Graph.t struct.

The “adjacent” field is a MapSet.t of ids of nodes that are adjacent to the node represented by the struct. An id in the adjacent set always represents a directed edge from the node to the adjacent node, regardless of whether the graph is directed or undirected (see the Graph module for more info). Conceptually, this represents a node’s adjacency list.

The “data” associates some arbitrary data with the node. Thus, the type of value stored in this field can be chosen to suit whatever purpose the node/graph is applied to.

The Node.t struct fields are considered private; do not access them directly, but rather use the interface functions provided.

Types

adjacency_list()

adjacency_list() :: MapSet.t(id())

id()

id() :: binary() | integer()

t()

t() :: %Graph.Node{adjacent: adjacency_list(), data: term(), id: id()}

t(struct)

t(struct) :: %Graph.Node{adjacent: adjacency_list(), data: struct, id: id()}

Functions

add_adjacent(node, adjNodeID)

Adds a node id to a node’s adjacency list, and returns the updated node.

get_data(node)

Returns the value of the node’s data field.

get_id(node)

Returns the id of a node.

has_adj_node?(node, adjNodeID)

Returns a boolean value indicating whether or not the node has the specified node id in its adjacency list.

new(id)

Returns a new Node struct with id.

new(id, data)

Returns a new Node struct with id and data.

remove_adjacent(node, adjNodeID)

Removes specified adjacent node from a node’s adjacency list, and returns the updated node.

If the specified adjacent node is not in the node’s adjacency list, prints an error message and returns nil.

set_data(node, data)

Sets the node’s data field to the specified value and returns the modified node.

set_id(node, id)

Sets the id of a node to the specified id, returning the updated node.